+++ AzureCommunicationUI/AzureCommunicationUI/Presentation/SwiftUI/Setup/SetupView.swift
@@ -49,7 +49,7 @@ struct SetupView: View {
                 JoiningCallActivityView(viewModel: viewModel.joiningCallActivityViewModel)
             } else {
                 PrimaryButton(viewModel: viewModel.joinCallButtonViewModel)
-                    .accessibility(identifier: LocalizationKey.joinCallAccessibilityLabel.rawValue)
+                    .accessibilityIdentifier(LocalizationKey.joinCallAccessibilityLabel.rawValue)
             }
         }
     }