+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.0.0-beta.3 (Upcoming)
 ### New Features
 - Updated minimum iOS version to 14.0. [#103](https://github.com/Azure/communication-ui-library-ios/pull/103)
+- Updated leave call confirmation drawer. [#129](https://github.com/Azure/communication-ui-library-ios/pull/129)
 - Implemented new local participant avatar parameter into launch method. [#131](https://github.com/Azure/communication-ui-library-ios/pull/131)
 
 ### Bugs Fixed